Changed: Add attachments directory to .gitignore. Changed: More attachment improvements using FineUploader.js. Lots and lots of changes here. Changed: PM Export now uses built-in PHP zip support. Probably needs PHP 5.3.x.

+++ forum/install/upgrade.php
@@ -45,6 +45,237 @@ if (!($forum_prefix_array = install_get_table_data())) {
     return;
 }
 
+$sql = "ALTER TABLE POST_ATTACHMENT_FILES CHANGE AID AID_OLD VARCHAR(32) NOT NULL, CHANGE ID ID MEDIUMINT(8) UNSIGNED NOT NULL, DROP PRIMARY KEY";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_FILES ADD COLUMN AID MEDIUMINT(8) UNSIGNED NOT NULL AUTO_INCREMENT FIRST, ADD PRIMARY KEY(AID)";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_FILES ADD COLUMN FILESIZE BIGINT(8) UNSIGNED NOT NULL AFTER MIMETYPE";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_FILES ADD COLUMN THUMBNAIL CHAR(1) DEFAULT 'N' NOT NULL AFTER FILESIZE";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_FILES CHANGE UID UID MEDIUMINT(8) UNSIGNED NOT NULL";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_FILES CHANGE FILENAME FILENAME VARCHAR(255) NOT NULL";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_FILES CHANGE MIMETYPE MIMETYPE VARCHAR(255) NOT NULL";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_FILES CHANGE HASH HASH VARCHAR(32) NOT NULL";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_IDS CHANGE AID AID_OLD CHAR(32) NOT NULL";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_IDS ADD COLUMN AID MEDIUMINT(8) UNSIGNED NOT NULL AFTER PID";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_IDS CHANGE FID FID MEDIUMINT(8) UNSIGNED NOT NULL";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_IDS CHANGE TID TID MEDIUMINT(8) UNSIGNED NOT NULL";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_IDS CHANGE PID PID MEDIUMINT(8) UNSIGNED NOT NULL";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_IDS DROP PRIMARY KEY";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_IDS ADD PRIMARY KEY(FID, TID, PID, AID)";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE PM_ATTACHMENT_IDS CHANGE AID AID_OLD CHAR(32) NOT NULL";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE PM_ATTACHMENT_IDS ADD COLUMN AID MEDIUMINT(8) UNSIGNED NOT NULL AFTER MID";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE PM_ATTACHMENT_IDS CHANGE MID MID MEDIUMINT(8) UNSIGNED NOT NULL";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE PM_ATTACHMENT_IDS DROP PRIMARY KEY";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE PM_ATTACHMENT_IDS ADD PRIMARY KEY(MID, AID)";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "REPLACE INTO POST_ATTACHMENT_IDS (FID, TID, PID, AID) ";
+$sql.= "SELECT POST_ATTACHMENT_IDS.FID, POST_ATTACHMENT_IDS.TID, ";
+$sql.= "POST_ATTACHMENT_IDS.PID, POST_ATTACHMENT_FILES.AID ";
+$sql.= "FROM POST_ATTACHMENT_FILES INNER JOIN POST_ATTACHMENT_IDS ";
+$sql.= "ON (POST_ATTACHMENT_IDS.AID_OLD = POST_ATTACHMENT_FILES.AID_OLD)";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_IDS DROP COLUMN AID_OLD";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "DELETE FROM POST_ATTACHMENT_IDS WHERE AID = 0";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "REPLACE INTO PM_ATTACHMENT_IDS (MID, AID) ";
+$sql.= "SELECT PM_ATTACHMENT_IDS.MID, POST_ATTACHMENT_FILES.AID ";
+$sql.= "FROM POST_ATTACHMENT_FILES INNER JOIN PM_ATTACHMENT_IDS ";
+$sql.= "ON (PM_ATTACHMENT_IDS.AID_OLD = POST_ATTACHMENT_FILES.AID_OLD)";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE PM_ATTACHMENT_IDS DROP COLUMN AID_OLD";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "DELETE FROM PM_ATTACHMENT_IDS WHERE AID = 0";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_FILES DROP COLUMN AID_OLD";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE POST_ATTACHMENT_FILES DROP COLUMN ID";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
 foreach ($forum_prefix_array as $forum_fid => $table_data) {
 
     if (!install_check_column_type($table_data['DATABASE_NAME'], "{$table_data['WEBTAG']}_ADMIN_LOG", "ENTRY", "longblob")) {
@@ -769,6 +1000,118 @@ foreach ($forum_prefix_array as $forum_fid => $table_data) {
         $valid = false;
         return;
     }
+
+    $sql = "UPDATE `{$table_data['PREFIX']}USER_PREFS` INNER JOIN POST_ATTACHMENT_FILES ";
+    $sql.= "ON (POST_ATTACHMENT_FILES.HASH = `{$table_data['PREFIX']}USER_PREFS`.AVATAR_AID) ";
+    $sql.= "SET `{$table_data['PREFIX']}USER_PREFS`.AVATAR_AID = POST_ATTACHMENT_FILES.AID";
+
+    if (!$result = $db->query($sql)) {
+
+        $valid = false;
+        return;
+    }
+
+    $sql = "UPDATE `{$table_data['PREFIX']}USER_PREFS` INNER JOIN POST_ATTACHMENT_FILES ";
+    $sql.= "ON (POST_ATTACHMENT_FILES.HASH = `{$table_data['PREFIX']}USER_PREFS`.PIC_AID) ";
+    $sql.= "SET `{$table_data['PREFIX']}USER_PREFS`.PIC_AID = POST_ATTACHMENT_FILES.AID";
+
+    if (!$result = $db->query($sql)) {
+
+        $valid = false;
+        return;
+    }
+
+    $sql = "ALTER TABLE `{$table_data['PREFIX']}USER_PREFS` CHANGE PIC_AID PIC_AID MEDIUMINT(11) NULL";
+
+    if (!$result = $db->query($sql)) {
+
+        $valid = false;
+        return;
+    }
+
+    $sql = "ALTER TABLE `{$table_data['PREFIX']}USER_PREFS` CHANGE AVATAR_AID AVATAR_AID MEDIUMINT(11) NULL";
+
+    if (!$result = $db->query($sql)) {
+
+        $valid = false;
+        return;
+    }
+
+    $sql = "UPDATE `{$table_data['PREFIX']}USER_PREFS` SET AVATAR_AID = NULL WHERE AVATAR_AID = 0";
+
+    if (!$result = $db->query($sql)) {
+
+        $valid = false;
+        return;
+    }
+
+    $sql = "UPDATE `{$table_data['PREFIX']}USER_PREFS` SET PIC_AID = NULL WHERE PIC_AID = 0";
+
+    if (!$result = $db->query($sql)) {
+
+        $valid = false;
+        return;
+    }
+}
+
+$sql = "UPDATE USER_PREFS INNER JOIN POST_ATTACHMENT_FILES ";
+$sql.= "ON (POST_ATTACHMENT_FILES.HASH = USER_PREFS.AVATAR_AID) ";
+$sql.= "SET USER_PREFS.AVATAR_AID = POST_ATTACHMENT_FILES.AID";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "UPDATE USER_PREFS INNER JOIN POST_ATTACHMENT_FILES ";
+$sql.= "ON (POST_ATTACHMENT_FILES.HASH = USER_PREFS.PIC_AID) ";
+$sql.= "SET USER_PREFS.PIC_AID = POST_ATTACHMENT_FILES.AID";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE USER_PREFS CHANGE PIC_AID PIC_AID MEDIUMINT(11) NULL";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "ALTER TABLE USER_PREFS CHANGE AVATAR_AID AVATAR_AID MEDIUMINT(11) NULL";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "UPDATE USER_PREFS SET AVATAR_AID = NULL WHERE AVATAR_AID = 0";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "UPDATE USER_PREFS SET PIC_AID = NULL WHERE PIC_AID = 0";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
+}
+
+$sql = "UPDATE USER_PREFS SET POST_PAGE = 63";
+
+if (!$result = $db->query($sql)) {
+
+    $valid = false;
+    return;
 }
 
 $sql = "DROP TABLE IF EXISTS SFS_CACHE";
@@ -971,7 +1314,7 @@ if (!$result = $db->query($sql)) {
     return;
 }
 
-$sql = "ALTER TABLE `USER` CHANGE `IPADDRESS` `IPADDRESS` VARCHAR(255) NOT NULL";
+$sql = "ALTER TABLE USER CHANGE IPADDRESS IPADDRESS VARCHAR(255) NOT NULL";
 
 if (!$result = $db->query($sql)) {
 
@@ -1065,6 +1408,68 @@ if (!$result = $db->query($sql)) {
 
     $valid = false;
     return;
+}
+
+if (($attachment_dir = forum_get_global_setting('attachment_dir'))) {
+
+    $attachment_dir = rtrim($attachment_dir, '/');
+
+    if (!($attachment_realpath = realpath($attachment_dir))) {
+
+        if (!($attachment_realpath = realpath(sprintf('../%s', $attachment_dir)))) {
+
+            $error_html.= "<h2>Could not locate attachment directory!</h2>\n";
+            $valid = false;
+            return;
+        }
+    }
+
+    $attachments = glob(sprintf('%s/*', $attachment_realpath));
+
+    $pattern_match = sprintf('/^%s\/([A-Fa-f0-9]{32})$/Du', preg_quote($attachment_realpath, '/'));
+
+    foreach ($attachments as $attachment) {
+
+        if (!preg_match($pattern_match, $attachment, $matches_array)) {
+            continue;
+        }
+
+        $hash = $db->escape($matches_array[1]);
+
+        $filesize = $db->escape(filesize($attachment));
+
+        $sql = "UPDATE POST_ATTACHMENT_FILES SET FILESIZE='$filesize' WHERE HASH='$hash'";
+
+        if (!$result = $db->query($sql)) {
+
+            $valid = false;
+            return;
+        }
+    }
+
+    $sql = "DELETE FROM POST_ATTACHMENT_FILES WHERE FILESIZE = 0";
+
+    if (!$result = $db->query($sql)) {
+
+        $valid = false;
+        return;
+    }
+
+    $sql = "DELETE FROM POST_ATTACHMENT_IDS WHERE AID NOT IN (SELECT AID FROM POST_ATTACHMENT_FILES)";
+
+    if (!$result = $db->query($sql)) {
+
+        $valid = false;
+        return;
+    }
+
+    $sql = "DELETE FROM PM_ATTACHMENT_IDS WHERE AID NOT IN (SELECT AID FROM POST_ATTACHMENT_FILES)";
+
+    if (!$result = $db->query($sql)) {
+
+        $valid = false;
+        return;
+    }
 }
 
 ?>
